Reorganizing gp bases and priors.
